Add XML Schema Definition references for 3.7/3.8.
I made those additions only in the section on the configuration file
in the assumption that this is where people will pick an initial XML
configuration template. I did not want to make the other places look
more complex than they have to be, espacially as the XSD is not
necessary for running PHPUnit.

I also did not elaborate on using the XSD or its potential benefits
to avoid the need of additional translations (haven't written any
Japanese or Chinese for a while).